divide by section
