+++ app/src/main/java/com/example/westudy/GeographyFragment.java
@@ -1,0 +1,27 @@
+package com.example.westudy;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+public class GeographyFragment extends Fragment {
+
+
+
+    public GeographyFragment() {
+        // Required empty public constructor
+    }
+
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_geography, container, false);
+    }
+}
